requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jsonapi==2.7.0
Faker==1.0.7
gunicorn==19.9.0
inflection==0.3.1
lxml==4.3.3
more-itertools==7.0.0
pluggy==0.11.0
psycopg2==2.8.2
py==1.8.0
pytest==4.5.0
python-dateutil==2.8.0
pytz==2019.1
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.3.3
sqlparse==0.3.0
text-unidecode==1.2
tqdm==4.32.1
urllib3==1.25.2
wcwidth==0.1.7
